==> python/3_clustering_for_manual_annotation.ipynb <==
%env PYTHONHASHSEED=0

## Loading the packages

import numpy as np
import pandas as pd
import os

import scanpy as sc
from anndata import AnnData, concat
from matplotlib.transforms import Bbox

# from scvi.model.utils import mde

import scvi 
from matplotlib import pyplot as plt
import seaborn as sns

from scipy import sparse
import gc
# from upsetplot import plot as UpSet
from itertools import product
import networkx as nx

from scipy.cluster import hierarchy

import re
import subprocess

import h5py

import itertools

# from scnym.api import scnym_api

# import pymn

from collections import Counter

from scvi.model.utils import mde


out_dir = "/home/anna/ClusterProjects/SA_bone_marrow/scfiles/"

# import tensorflow as tf
# import tensorboard as tb
# tf.io.gfile = tb.compat.tensorflow_stub.io.gfile

pd.set_option('display.max_columns', 500)

### Directory TM_combined_datasets_SA_ann_obs
# combined_datasets = sc.read_h5ad(filename = "../data/h5ad/TM_combined_datasets_SA_ann.h5ad")
# obs_plk = pd.read_pickle("../data/h5ad/TM_combined_datasets_SA_ann_obs.plk")
# combined_datasets.obs = obs_plk

hvg = sc.read("../data/h5ad/hvg_integrated_170824.h5ad")

hvg.obsm

### Extracting only SA experiment from the scNym datasets

# SA_exps = combined_datasets[combined_datasets.obs["study_id"].isin(['SA_pilot', 'SA_final'])]

### Reading SA experiment

SA_exps = sc.read_h5ad(filename = out_dir + "SA_exp_scNym_annotated.h5ad")
obs_plk = pd.read_pickle(out_dir + "SA_exp_scNym_annotated.plk")
SA_exps.obs = obs_plk

hvg_cells = hvg.obs_names
SA_exp_names = SA_exps.obs_names

SA_hvg_cells = [x for x in SA_exp_names if x[:-2] in hvg_cells]

SA_cellname_counts = Counter([x[:-2] for x in SA_hvg_cells])
SA_unique_names = [x for x in SA_hvg_cells if SA_cellname_counts[x[:-2]] == 1]

SA_exps_subset = SA_exps[SA_unique_names]

SA_exps_subset.obs_names = [x[:-2] for x in SA_exps_subset.obs_names]

hvg_subset = hvg[SA_exps_subset.obs_names]

hvg_subset.obs["scNym_condensed"] = SA_exps_subset.obs["scNym_condensed"]
hvg_subset.obs["scNym_condensed_confidence"] = SA_exps_subset.obs["scNym_condensed_confidence"]

SA_exps_HSC = SA_exps_subset[SA_exps_subset.obs["scNym_condensed"] == "hematopoietic precursor cell"]
SA_exps_mature = SA_exps_subset[SA_exps_subset.obs["scNym_condensed"] != "hematopoietic precursor cell"]

### HSC population

hvg_HSC_v2 = hvg_subset[SA_exps_HSC.obs_names.intersection(hvg_subset.obs_names)]
# SA_HSC_v2 = SA_exps_HSC[SA_exps_HSC.obs_names.intersection(hvg_subset.obs_names)]

hvg_HSC_v2.shape

hvg_HSC_v2.layers

### Loading the TM HSC SA annotated dataset

combined_HSC = sc.read_loom(filename = "../data/loom/TM_combined_HSC_SA_ann.loom")

annotated_SA_HSC = combined_HSC[ combined_HSC.obs["study_id"].isin(["SA_final", "SA_pilot"])]
annotated_SA_HSC.obs_names = annotated_SA_HSC.obs["obs_names"]

del(combined_HSC)

ann_counts = Counter([x[:-4] for x in annotated_SA_HSC.obs_names])
unique_names = [x for x in annotated_SA_HSC.obs_names if ann_counts[x[:-4]] == 1]
annotated_SA_HSC = annotated_SA_HSC[unique_names]
annotated_SA_HSC.obs_names = [x[:-4] for x in annotated_SA_HSC.obs_names]
common_cells = annotated_SA_HSC.obs_names.intersection(hvg_HSC_v2.obs_names)

Counter(hvg_HSC_v2.obs["scNym_HPC"])

hvg_HSC_v2.obs["scNym_HPC"] = annotated_SA_HSC[common_cells].obs["scNym_HPC"]

sum(common_cells == hvg_HSC_v2.obs_names)

# sc.pp.normalize_total(hvg_HSC_v2, target_sum=1e6)
# sc.pp.log1p(hvg_HSC_v2)

hvg_HSC_v2.shape

hvg_HSC_v2.layers

hvg_HSC_v2.obsm["X_scVI"]

hvg_HSC_v2.layers

### Clustering for HSC

sc.tl.pca(hvg_HSC_v2, svd_solver='arpack', layer = "scVI_normalized")
sc.pp.neighbors(hvg_HSC_v2, n_neighbors=10, n_pcs=40, use_rep = "X_pca")
sc.tl.umap(hvg_HSC_v2, random_state=4567898)
sc.tl.leiden(hvg_HSC_v2, resolution=0.25, 
             n_iterations=100)

hvg_HSC_v2.obs

sc.set_figure_params(figsize = (8,6))
sc.pl.embedding(
    hvg_HSC_v2,
    basis="umap",
    # color=['Mariem_annotation'],
    color=['leiden', 'batch', "scNym_HPC"],
    frameon=False,
    ncols=1,
    palette = "tab20",
    save = "hvg_scVI_HSC_v2_leiden_batch_scNym_HPC.pdf"
)

hvg_HSC_v2.obsm["X_mde"] = mde(hvg_HSC_v2.obsm["X_scVI"])

sc.pl.embedding(
    hvg_HSC_v2,
    basis="X_mde",
    color=['leiden', 'batch', "scNym_HPC"],
    frameon=False,
    ncols=1,
)

sc.set_figure_params(figsize = (8,6))
sc.pl.embedding(
    hvg_HSC_v2,
    basis="umap",
    # color=['Mariem_annotation'],
    color=['leiden', 'batch', "scNym_HPC"],
    frameon=False,
    ncols=1,
    palette = "tab20",
    save = "hvg_scVI_HSC_v2_leiden_manual_annotation_legend.pdf"
)

# SA_HSC_v2.write_loom(filename = out_dir + "SA_HSC_v2_2508.loom")
hvg_HSC_v2.write_loom(filename = out_dir + "hvg_HSC_v2_2608.loom")

### Clustering for mature population 

hvg_mature_v2 = hvg_subset[SA_exps_mature.obs_names.intersection(hvg_subset.obs_names)]
# SA_mature_v2 = SA_exps_mature[SA_exps_mature.obs_names.intersection(hvg_subset.obs_names)]

del hvg_mature_v2.obsm

hvg_mature_v2.layers

# sc.pp.normalize_total(hvg_mature_v2, target_sum=1e6)
# sc.pp.log1p(hvg_mature_v2)
sc.tl.pca(hvg_mature_v2, layer="scVI_normalized")
sc.pp.neighbors(hvg_mature_v2, use_rep="X_pca")
sc.tl.umap(hvg_mature_v2, random_state=46468856)
sc.tl.leiden(hvg_mature_v2, n_iterations=-1, resolution=0.5)

sc.set_figure_params(figsize = (8,6))
sc.pl.embedding(
    hvg_mature_v2,
    basis="umap",
    # color=['Mariem_annotation'],
    color=['leiden', 'batch'],
    frameon=False,
    ncols=1,
    palette = "tab20",
    save = "hvg_scVI_HSC_v2_leiden_manual_annotation_legend.pdf"
)

hvg_mature_v2.write_loom(filename = out_dir + "hvg_mature_v2_2608.loom")
