==> python/4_dynamo_cell_cycle.ipynb <==
import scanpy as sc
# import scvi
import glob
import os
from functools import reduce
import anndata as ad
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns
import pandas as pd
from scipy.io import mmread
import gzip
import dynamo as dyn
import velocyto as vcy

### Reading in loom from velocyto

vlm_pilot = vcy.VelocytoLoom("../velocyto_run//output/MR_30_10X_SA_PBS_BM_22112021_transcriptome_possorted_genome_bam_IFOE5.loom")
vlm_final = vcy.VelocytoLoom("../velocyto_run/output/MR_33_10X_SA_PBS_BM_transcriptome_possorted_genome_bam_IJA5R.loom")

### Configuring dynamo

dyn.configuration.set_figure_params('dynamo', background='white') # jupter notebooks
dyn.configuration.set_figure_params('dynamo', background='black') # presentation
# dyn.configuration.set_pub_style() # manuscript

# bm = sc.read("../output/bm_integrated.h5ad")
hvg = sc.read("../data/h5ad/hvg_integrated_170824.h5ad")

hvg.obs['pMito'] =hvg.obs['percent_mt']

hvg.obsm['X_scVI']

dyn.pp.recipe_monocle(hvg)

hvg.layers

hvg

type(hvg)

del(hvg.uns["cell_phase_genes"])

hvg.write_h5ad("../data/h5ad/hvg_integrated_170824_cell_cycle_phases.h5ad")
